requirements.txt
pandas==2.0.0
Pillow==9.5.0
streamlit==1.10.0
